Mapping Integral Constant to Types
This is very useful when we need to discriminate in compile time.
E.g. If a object could or could not be polimorphic and we need a different behaviour in each case.
With a if-else the code will not compile.

A template with different ints parameter is taken as different types. This can use to create overloaded functions that are called in the different cases, and the compile will accept it.