Stripping the const qualifier of a type
We could want a template to be able to take const parameter to the template.
But then, inside the template class or template function objects created
with such types will be const.

Maybe we dont want this. E.g. shared_ptr<const T>
Inside the template class or function we can use this technique to strip
the const qualifier